Fixed transfer queue not having a lock around its command pool and fixed a few warnings of precision loss (well, more so made them go away)

+++ src/rendering/Buffer.cpp
@@ -37,7 +37,7 @@ void AlignedMemcpy(uint8_t bytes, VkDeviceSize destMemAlignment, void* src, void
 			(char*)src + src_offset,
 			sizeof(bytes));
 		src_offset += 1;
-		dest_offset += destMemAlignment;
+		dest_offset += (int)destMemAlignment;
 	
 	}
 }
